+++ TestProject/app/src/main/java/net/aquadc/mike/plugin/test/interfaces.kt
@@ -9,7 +9,22 @@ class Iii : AbstractList<Nothing>(), Runnable, Cloneable, () -> Unit {
         top(1, 1, emptyList(), this, this) { 0 }
         1.topExt(1, emptyList(), this, this, this)
         member(1, 1, emptyList(), this, this, this)
-        1.memberExt(1, emptyList(), this, r = this, cl = this)
+        member(
+            ser = 1,
+            f = this,
+            cmp = 1,
+            r = this,
+            cl = this,
+            lst = emptyList(),
+        )
+        1.memberExt(1, emptyList(), this, this, this)
+        1.memberExt(
+            r = this,
+            cl = this,
+            cmp = 1,
+            func = this,
+            lst = emptyList(),
+        )
         Toast.makeText(null, null, Toast.LENGTH_LONG).show()
     }
 
@@ -26,9 +41,9 @@ class Iii : AbstractList<Nothing>(), Runnable, Cloneable, () -> Unit {
     override fun invoke() {
     }
 
-    private fun member(a: Serializable, b: Comparable<*>, c: List<Number>, f: () -> Unit, cl: Cloneable, r: Runnable) {
+    private fun member(ser: Serializable, cmp: Comparable<*>, lst: List<Number>, f: () -> Unit, cl: Cloneable, r: Runnable) {
     }
-    private fun Serializable.memberExt(b: Comparable<*>, c: List<Number>, f: () -> Unit, cl: Cloneable, r: Runnable) {
+    private fun Serializable.memberExt(cmp: Comparable<*>, lst: List<Number>, func: () -> Unit, cl: Cloneable, r: Runnable) {
     }
 }